% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alias2SymbolUsingNCBI}
\alias{alias2SymbolUsingNCBI}
\title{Convert Gene Aliases to Official Gene Symbols}
\usage{
alias2SymbolUsingNCBI(
  alias,
  gene.info.file,
  required.columns = c("GeneID", "Symbol", "description")
)
}
\arguments{
\item{alias}{character vector of gene aliases}

\item{gene.info.file}{the name of a gene information file downloaded from the
NCBI.}

\item{required.columns}{character vector of columns from the gene information
file that are required in the output.}
}
\value{
A data frame with the latest gene symbols
}
\description{
Maps gene alias names to official gene symbols.
}
\examples{
library(labyrinth)
\donttest{update_gene_symbol('IL-6')}

}
