% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.dgCMatrix.R
\name{assert_dgCMatrix}
\alias{assert_dgCMatrix}
\title{Check if the sparse matrix valid}
\usage{
assert_dgCMatrix(adj_matrix)
}
\arguments{
\item{adj_mat}{The sparse matrix to check}
}
\description{
Check if the sparse matrix valid
}
