% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clearDomain}
\alias{clearDomain}
\title{Clear Neo4j sub-graph}
\usage{
clearDomain(domain)
}
\arguments{
\item{domain}{A string of which is attached to nodes and relationships as the domain property in the database}
}
\description{
Clear sub-graph in Neo4j with given domain.
}
\examples{
clearDomain('karate')
}

