% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disease_ids}
\alias{disease_ids}
\title{Disease MeSH Identifiers}
\format{
A character vector.
}
\source{
The disease MeSH identifiers were obtained from the National Library of
  Medicine's Medical Subject Headings (MeSH) database 
  (https://www.nlm.nih.gov/mesh/).
}
\usage{
disease_ids
}
\description{
A character vector containing Medical Subject Headings (MeSH) identifiers for
  all human diseases.
}
\details{
The `disease_ids` object is a character vector that contains MeSH identifiers
  (MeSH IDs) for various diseases. Each element in the vector represents a
  unique MeSH identifier, which can be used to identify and reference
  specific diseases in the MeSH taxonomy.

MeSH identifiers are hierarchically organized and provide a standardized way
  of representing and categorizing diseases and other medical concepts.
}
\examples{
data(disease_ids, package = "labyrinth")
length(disease_ids)

# Check if a specific MeSH identifier is present
"D000544" \%in\% disease_ids  # Alzheimer's disease
}
\keyword{datasets}
