% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_disease_genes.R
\name{disease_impact_score}
\alias{disease_impact_score}
\title{Get disease impact score}
\usage{
disease_impact_score(expr, control, max.iter = 10000)
}
\arguments{
\item{expr}{The expression matrix}

\item{control}{A vector or scalar of the index of the normal samples.}

\item{max.iter}{Maximal number of iterations of the augumented Lagrange
multiplier algorithm, which is used in robust PCA algorithm.}
}
\value{
A vector of the disease impact score
}
\description{
Get disease impact score
}
\examples{
# no examples
# TODO: fill the examples

}
