% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_annot}
\alias{drug_annot}
\title{Drug annotation dataset}
\format{
A data.table with the following columns:
\describe{
  \item{drug_id}{A character vector representing the drug identifier.}
  \item{drug_name}{A character vector representing the name of the drug.}
}
}
\source{
The drug annotations were obtained from ChEMBL, DrugBank, and CTD.
}
\usage{
drug_annot
}
\description{
A data.table containing drug identifiers and their corresponding names.
}
\details{
The `drug_annot` data set provides a mapping between drug identifiers and
  their corresponding names. Each row represents a unique drug identifier
  (`drug_id`) and its associated name (`drug_name`).

The drug identifiers and names in this data set are sourced from ChEMBL,
  DrugBank, and CTD to extract synonyms for the same drug.
}
\examples{
library(labyrinth)

data(drug_annot, package = "labyrinth")

# Access drug names for specific identifiers
drug_annot[drug_annot$drug_id == "2838", ]$drug_name  # Imatinib

# Count the number of unique drugs
length(unique(drug_annot$drug_id))

# Visualize the data
drug_id <- drug_annot$drug_id
hist(table(drug_id))
}
\keyword{datasets}
