% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_disease}
\alias{gene_disease}
\title{Gene-disease associations}
\format{
A \code{\link[methods:namedList-class]{named list}}, where each element is a
  character vector representing the genes associated with a specific MeSH
  disease identifier (MeSH ID). The names of the list elements correspond to
  the MeSH IDs.
}
\source{
The gene-disease associations were obtained from the DisGeNET database
  (https://www.disgenet.org/), which integrates information from various 
  sources to provide gene-disease associations.
}
\usage{
gene_disease
}
\description{
A list containing gene-disease associations obtained from DisGeNET with the
  confidence level of 60%.
}
\details{
The `gene_disease` data set contains gene-disease associations extracted
  from the DisGeNET database with a confidence level of 60%. Each element of
  the list represents a MeSH disease identifier, and the corresponding
  character vector contains the gene identifiers associated with that
  disease.
}
\examples{
data(gene_disease, package = "labyrinth")
gene_disease[["D000168"]]
}
\keyword{datasets}
