% Generated by roxygen2 (4.0.2): do not edit by hand
\name{globalStats}
\alias{globalStats}
\title{Graph level measures}
\usage{
globalStats(g, glob)
}
\arguments{
\item{g}{An igraph object}

\item{glob}{A list of which netowrk parameters should be calculated}
}
\description{
Function to compute a set of measures for the entire network.
}
\details{
Will add to the parameteres into a dataframe.
}
\examples{
require(igraph)
require(igraphdata)

data(karate)

globalStats(karate, c("Degree", "Radius", 'Girth'))
}

