% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.dgCMatrix.R
\name{is.dgCMatrix}
\alias{is.dgCMatrix}
\title{Check if a matrix is `dgCMatrix` class}
\usage{
is.dgCMatrix(mat)
}
\arguments{
\item{mat}{The matrix to check}
}
\value{
A boolean variable
}
\description{
Check if a matrix is `dgCMatrix` class
}
