% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labyrinth-package.R
\name{labyrinth-package}
\alias{labyrinth-package}
\alias{labyrinth}
\title{Labyrinth: A Simulation of Knowledge Networks}
\description{
One paragraph description of what the package does
  as one or more full sentences. I can use Config/testthat/parallel as
  true to boost the test efficiency.
}
\details{
labyrinth
}
\examples{
# if (!"devtools" \%in\% as.data.frame(installed.packages())$Package)
#   install.packages("devtools")
# devtools::install_github("randef1ned/labyrinth")

}
