% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labyrinth_analysis.R
\name{LabyrinthAnalyze}
\alias{LabyrinthAnalyze}
\title{Produce a data frame with analysis information for the imputation}
\usage{
LabyrinthAnalyze(orig, masked, imputed)
}
\arguments{
\item{orig}{original (filtered) unmasked unimputed vcfR object}

\item{masked}{masked vcfR object}

\item{imputed}{imputed vcfR object}
}
\value{
data frame contatianing analysis information
}
\description{
This will return a data frame with one row for each masked site. The columns
of the data frame are `orig` which contains the number of alternate alleles
in the original call (0, 1, or 2), `imputed` which contains the number of '
alternate alleles in the imputed call (0, 1, or 2), `correct` as a logical '
indicating if these two calls are the same, `p.ref`, `p.het`, and `p.alt` '
indicating respectively the probability returned for each genetic state by '
LaByRInth. These three values will be NA if the imputed vcf does not
contiain posterior probablilities converted to phred scores.
}
\author{
Jason Vander Woude
}
