% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labyrinth.R
\name{LabyrinthCalcProgenyHet}
\alias{LabyrinthCalcProgenyHet}
\title{Calculate expected heterozygosity in a population}
\usage{
LabyrinthCalcProgenyHet(breed.scheme)
}
\arguments{
\item{breed.scheme}{Supported breeding scheme}
}
\value{
Expected heterozygousity proportion
}
\description{
Calculate expected heterozygosity in a population. Supported breeding
schemes are "F1BC1" and "F<x>" where <x> is replaced by any (possibly
multi-digit) number.
}
\author{
Jason Vander Woude
}
