% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labyrinth.R
\name{LabyrinthFilter}
\alias{LabyrinthFilter}
\title{Remove bad data prior to imputing}
\usage{
LabyrinthFilter(vcf, out.file, parents, require.hom.poly = FALSE)
}
\arguments{
\item{vcf}{File path or vcfR object to impute.}

\item{out.file}{File path of output file.}

\item{parents}{Character vector with names of the two parents.}

\item{require.hom.poly}{Logical indicating if the sites that are kept are
required to by homozygous within and polymorphic between (i.e. should
the parents be homozygous for different alleles at every kept site
according to the genotype (gt) field in the vcf). The implementation
of LB-Impute that LaByRInth is based on required this, but this
version of LaByRInth does not. This functionality is mostly included
for legacy purposes and should generally be set to false.}
}
\value{
A vcfR object with all sites removed that don't meet the specified
        criteria.
}
\description{
Remove all sites that are not biallelic and possibly sites which are not
homozygous within and polymorphic between.
}
\examples{
input <- system.file(
    "extdata",
    "vcf-files",
    "original-lakin-fuller-sample.vcf",
    package = "LaByRInth",
    mustWork = TRUE)
output <- tempfile(fileext="-filtered-result.vcf.gz")
print(output)
result <- LabyrinthFilter(
    vcf = input,
    out.file = output,
    parents = c("LAKIN", "FULLER"),
    require.hom.poly = TRUE)
}
\author{
Jason Vander Woude
}
