% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labyrinth.R
\name{LabyrinthImpute}
\alias{LabyrinthImpute}
\title{Run entire imputation including filtering and quality control}
\usage{
LabyrinthImpute(vcf, out.file, parents, breed.scheme, progeny.het = NULL,
  read.err = NULL, min.posterior, geno.err = 0.01, parent.het = 0.01,
  require.hom.poly = FALSE, parallel = TRUE, cores = 4)
}
\arguments{
\item{vcf}{File path or vcfR object to impute.}

\item{out.file}{File path of output file.}

\item{parents}{Character vector with names of the two parents.}

\item{breed.scheme}{Population type (e.g. F2, F5, F1BC1).}

\item{min.posterior}{Numeric specifying the minimum probability for any call
that is kept.}

\item{geno.err}{Estimate of proportion of calls genotyped incorrectly.}

\item{parent.het}{Estimate of proportion of sites in each parent that are
truly heterozygous.}

\item{require.hom.poly}{Logical indicating if the sites that are kept are
required to by homozygous within and polymorphic between (i.e. should
the parents be homozygous for different alleles at every kept site
according to the genotype (gt) field in the vcf). The implementation
of LB-Impute that LaByRInth is based on required this, but this
version of LaByRInth does not. This functionality is mostly included
for legacy purposes and should generally be set to false.}

\item{parallel}{Logical indicating if imputation should be run in parallel
or serial.}

\item{cores}{Numeric indicating how many sub-processes should be spawned if
running in parallel.}
}
\value{
A vcfR object with all sites removed that don't meet the filtering
        criteria, both parents and progeny imputed and low probability sites
        removed.
}
\description{
Run all four other functions (LabyrinthFilter, LabyrinthImputeParents,
LabyrinthImputeProgeny, and LabyrinthQualityControl) in sequence and save
the intermediate files to a temporary directory. Running this function will
produce equivalent results as running all four functions separately; the
only difference will be where the files are saved. This is provided simply
for user convenience.
}
\examples{
input <- system.file(
    "extdata",
    "vcf-files",
    "original-lakin-fuller-sample.vcf",
    package = "labyrinth",
    mustWork = TRUE)
output <- tempfile(fileext="-result.vcf.gz")
result <- LabyrinthImpute(
    vcf = input,
    out.file = output,
    parents = c("LAKIN", "FULLER"),
    breed.scheme = "F5",
    progeny.het = 0.5^4, # expected heterozygosity of F5 population
    min.posterior = 0.8,
    geno.err = 0.015,
    parent.het = 0.005,
    require.hom.poly = TRUE,
    parallel = FALSE,
    cores = 1)
}
\author{
Jason Vander Woude
}
