% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labyrinth.R
\name{LabyrinthImputeProgeny}
\alias{LabyrinthImputeProgeny}
\title{Offspring imputation}
\usage{
LabyrinthImputeProgeny(parental, out.file, parallel = TRUE, cores = 4)
}
\arguments{
\item{parental}{An object of class parental.imputation (a result of the
LabyrinthParentalImputation function) which will contain information
about the population including the parental imputation information.}

\item{out.file}{File path of output file.}

\item{parallel}{Logical indicating if imputation should be run in parallel
or serial.}

\item{cores}{Numeric indicating how many sub-processes should be spawned if
running in parallel.}
}
\value{
A vcfR object with both parents and progeny imputed.
}
\description{
Imputes all of the progeny (offspring).
}
\examples{
input <- system.file(
    "extdata",
    "vcf-files",
    "parental-lakin-fuller-sample.rds",
    package = "labyrinth",
    mustWork = TRUE)
output <- tempfile(fileext="-progeny-result.vcf.gz")
print(output)
result <- LabyrinthImputeProgeny(
    parental = input,
    out.file = output,
    parallel = FALSE,
    cores = 1)
}
\author{
Jason Vander Woude
}
