% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_replication.R
\name{LabyrinthMask}
\alias{LabyrinthMask}
\title{Mask the most confident calls in a VCF file for validation}
\usage{
LabyrinthMask(vcf, parents, out.file, mask.prop, top.prop, rerr = 0.05)
}
\arguments{
\item{vcf}{File path or vcfR object to impute.}

\item{parents}{Character vector with names of the two parents.}

\item{out.file}{File path of output file.}

\item{rerr}{Estimated read error rate of the sequencing process (used in
determining the likelihood of the reads).}

\item{depth}{Minimum read depth of masked sites.}

\item{lik.ratio}{Minimum likelihood ratio of masked sits.}
}
\value{
A vcfR object with all sites removed that meet the masking criteria.
}
\description{
Compute the liklihood of the the reads for each of the three genotypes
(homozygous reference, heterozygous, and homozygous alternate) and return
the ratio of the the greatest liklihood over the sum of the other
two. This gives an indication of likly it is that the inferred call is
the true genotype.
}
\author{
Jason Vander Woude
}
