% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labyrinth_analysis.R
\name{LabyrinthPlotPosteriors}
\alias{LabyrinthPlotPosteriors}
\title{Generate a plot of imputation metrics against posterior probabilities}
\usage{
LabyrinthPlotPosteriors(labyrinth.df, lb.impute.df, fsfhap.df,
  dataset.name, out.file = NULL, n = 21)
}
\arguments{
\item{labyrinth.df}{a data frame of the analysis of a LaByRInth imputation}

\item{lb.impute.df}{a data frame of the analysis of an LB-Impute imputation}

\item{fsfhap.df}{a data frame of the analysis of an FSF-Hap imputation}

\item{dataset.name}{single element character vector used in plot title}

\item{out.file}{file to save the plot (if NULL, plot is sent to standard graphics)}

\item{n}{how many different posterior thresholds to use}
}
\value{
nothing
}
\description{
Generate and possibly save a plot where the x-axis is the posterior
probability thresholds and the y-axis is various metrics indicating how good
the imputation is.
}
\author{
Jason Vander Woude
}
