% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_replication.R
\name{LabyrinthPreparePublicationData}
\alias{LabyrinthPreparePublicationData}
\title{Reproduce the prepared data of the LaByRInth publication}
\usage{
LabyrinthPreparePublicationData(dataset, output.dir)
}
\arguments{
\item{dataset}{one of the following strings: "Lakin-Fuller", "HincII",
"RsaI", "IBM-RIL", "Sim-A-F1BC1}

\item{output.dir}{a directory where all subdirectories and filtered and
masked files will be placed.}
}
\description{
This function will filter and mask one of the five data files from the
LaByRInth publication so that the results can be verified. The filtered and
masked result files will be placed in the specified output directory.
}
\author{
Jason Vander Woude
}
