% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data}
\alias{load_data}
\title{Load External Data}
\usage{
load_data(required_file)
}
\arguments{
\item{required_file}{Required file name without extensions from GitHub
\href{https://github.com/randef1ned/labyrinth/tree/master/extdata}{repo}.}
}
\value{
The content of the required data.
}
\description{
This function loads the external data to the R environment. If the required
  file is not found locally, it attempts to download the file from author's
  GitHub. In this package, it is used to download external data from
  \href{https://github.com/randef1ned/labyrinth/tree/master/extdata}{this}
  folder, such as the pre-trained model file, drug annotation file, etc.
}
\details{
The function first checks if the required file exists in the labyrinth
  directory (defined by \code{tools::R_user_dir('labyrinth')}). If the
  required file is not found, it attempts to download the file from GitHub
  repo up to five times. If the download fails after five attempts, an error
  is thrown. If the file is successfully downloaded or already exists
  locally, the function will return the file content.
}
\examples{
\dontrun{
# Load the trained model
model <- load_data("model")

# Load the drug annotation data
drug_annot <- load_data("drug_annot")
}
}
