% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_model}
\alias{load_model}
\title{Load Trained Model}
\usage{
load_model()
}
\value{
The loaded model object.
}
\description{
This function loads the pre-trained labyrinth model to the R environment. If 
  the model file is not found locally, it attempts to download the file from 
  author's GitHub.
}
\details{
The function first checks if the model file exists in the labyrinth directory
  (defined by \code{tools::R_user_dir('labyrinth')}). If the model file is 
  not found, it attempts to download the file from GitHub repo up to five 
  times. If the download fails after five attempts, an error is thrown. If 
  the file is successfully downloaded or already exists locally, trained 
  model file is loaded into the current environment.
}
\examples{
\donttest{
# Load the model
model <- load_model()
}
}
