% Generated by roxygen2 (4.0.2): do not edit by hand
\name{localStats}
\alias{localStats}
\title{Vertex level measures}
\usage{
localStats(g, stats)
}
\arguments{
\item{g}{An igraph object}

\item{stats}{A list of which netowrk parameters should be calculated}
}
\description{
Function to compute a set of measures for each vertex in a given network.
}
\details{
Will add to the parameteres into a dataframe with the vertex name
}
\examples{
require(igraph)
require(igraphdata)

data(karate)

localStats(karate, c("degree", "closeness", 'hub'))
}

