% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mazeWalker}
\alias{mazeWalker}
\title{Mazewalker}
\usage{
mazeWalker(constraints, stats)
}
\arguments{
\item{constraints}{A method pull a certain sub-graph}

\item{stats}{A list of which netowrk parameters should be calculated}
}
\description{
Utilize igraph to compute network statistics to write back to Neo4j
}
\details{
Will write network parameters back tp Neo4j
}
\examples{
require(igraph)
require(igraphdata)

data(karate)
ingest(karate, 'knows')

mazeWalker('karate', c("degree", "closeness", "betweeness", 'hub'))
}

