% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mesh_annot}
\alias{mesh_annot}
\title{MeSH Annotation Data}
\format{
A \code{\link[data.table:data.table]{data.table}} with the following 
  columns:
\describe{
  \item{mesh_id}{A character vector representing the MeSH identifier.}
  \item{mesh_term}{A character vector representing the MeSH term.}
  \item{group_id}{A character vector representing the group identifier obtained from NCBI.}
  \item{group_name}{A character vector representing the group name obtained from NCBI.}
}
}
\source{
The MeSH terms were obtained from the National Library of Medicine's Medical
  Subject Headings (MeSH) database (https://www.nlm.nih.gov/mesh/), and the
  group information was obtained from the National Center for Biotechnology
  Information (NCBI).
}
\usage{
mesh_annot
}
\description{
A data frame containing annotated Medical Subject Headings (MeSH) terms and
  their associated group information.
}
\examples{
data(mesh_annot, package = "labyrinth")
}
\keyword{datasets}
