% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mesh_hierarchy}
\alias{mesh_hierarchy}
\title{MeSH hierarchy network}
\format{
A \code{\link[data.table:data.table]{data.table}} with the following columns:
\describe{
  \item{from_id}{A character vector representing the MeSH identifier of the parent term.}
  \item{to_id}{A character vector representing the MeSH identifier of the child term.}
}
}
\source{
The MeSH hierarchy data was obtained from the National Library of Medicine's
  Medical Subject Headings (MeSH) database (https://www.nlm.nih.gov/mesh/).
}
\usage{
mesh_hierarchy
}
\description{
An edgelist network representing the hierarchical structure of the Medical
  Subject Headings (MeSH) terms.
}
\details{
The `mesh_hierarchy` data frame is an edgelist network, where each row
  represents a directed edge from a parent MeSH term (`from_id`) to a child
  MeSH term (`to_id`). This structure captures the hierarchical relationships
  between MeSH terms.
}
\examples{
data(mesh_hierarchy, package = "labyrinth")
}
\keyword{datasets}
