% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncbi_info}
\alias{ncbi_info}
\title{NCBI Info}
\format{
A \code{\link[base:data.frame]{data.frame}} representing the 
  information of NCBI genes.
}
\usage{
ncbi_info
}
\description{
TODO: enhance documentation
}
\examples{
library(labyrinth)

data("ncbi_info", package = "labyrinth")

}
\keyword{datasets}
