% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_gram.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function of Spreadgram}
\usage{
sigmoid(ax, ay, u = 1)
}
\arguments{
\item{ax}{A vector of the activation rates of x, which is a set of the y's
neighbors}

\item{ay}{A scalar of the activation rate of node y.}

\item{u}{If \eqn{x \in N(y)} then \eqn{u = 1}, otherwise \eqn{u = 0}.}
}
\value{
A vector of sigma
}
\description{
The Sigmoid function is one of the basic function in Spreadgram algorithm.
  This function calculates the Sigmoid function of a specific node given the
  activation rates of the nodes and its neighbors.
  The equation writes
  \deqn{\sigma (x,y) = {{\exp [a(x)a(y)]} \over {1 + \exp [a(x)a(y)]}}}.
}
\examples{

sigmoid(1:4, 2)

}
