% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_gene_symbol}
\alias{update_gene_symbol}
\title{Update gene symbol using NCBI Entrez Gene}
\usage{
update_gene_symbol(
  gene.pool,
  ncbi = NA,
  species = c("Homo_sapiens", "Mus_musculus", "Pan_troglodytes", "Rattus_norvegicus",
    "Archaea", "Bacteria", "Escherichia_coli_str._K-12_substr._MG1655",
    "Pseudomonas_aeruginosa_PAO1", "Ascomycota", "Microsporidia", "Bos_taurus",
    "Penicillium_rubens", "Saccharomyces_cerevisiae", "Anopheles_gambiae",
    "Caenorhabditis_elegans", "Drosophila_melanogaster", "Canis_familiaris",
    "Sus_scrofa", "Danio_rerio", "Gallus_gallus", "Xenopus_laevis", "Xenopus_tropicalis",
    "Arabidopsis_thaliana", "Chlamydomonas_reinhardtii", 
     "Oryza_sativa",
    "Zea_mays", "Plasmodium_falciparum", "Retroviridae")
)
}
\arguments{
\item{gene.pool}{The gene list}

\item{ncbi}{The location to the NCBI file directory}

\item{species}{Optional. To specify the species}
}
\value{
A vector with the latest gene symbols
}
\description{
Update gene symbol using NCBI Entrez Gene
}
\examples{
library(labyrinth)
\donttest{update_gene_symbol('IL-6')}
}
\seealso{
[alias2SymbolUsingNCBI()]
}
