% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_replication.R
\name{LabyrinthImputePublicationData}
\alias{LabyrinthImputePublicationData}
\title{Impute the masked publication data}
\usage{
LabyrinthImputePublicationData(dataset, output.dir, parallel = FALSE,
  cores = 1, mask.IDs = NULL)
}
\arguments{
\item{dataset}{one of the following strings: "Lakin-Fuller", "HincII",
"RsaI", "IBM-RIL", "Sim-A-F1BC1}

\item{output.dir}{a directory where all subdirectories and filtered and
masked files will be placed. MUST BE THE SAME AS THE output.dir
PARAMETER USED WHEN RUNNING LabyrinthPreparePublicationData.}

\item{parallel}{Logical indicating if imputation should be run in parallel
or serial.}

\item{cores}{Numeric indicating how many sub-processes should be spawned if
running in parallel.}
}
\description{
First run LabyrinthPreparePublicationData,
}
\author{
Jason Vander Woude
}
