% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncbi_info}
\alias{ncbi_info}
\title{NCBI Gene Information}
\format{
A \code{\link[data.table:data.table]{data.table}} with the following
  columns:

\describe{
  \item{type}{A character vector representing the type of species.}
  \item{species}{A character vector representing the species that NCBI supports in its online platform.}
}
}
\source{
This data was obtained from the National Center for Biotechnology Information
  (NCBI) Gene database (https://www.ncbi.nlm.nih.gov/gene).
}
\usage{
ncbi_info
}
\description{
A data table containing meta information about NCBI genes, including their
  type and associated species.
}
\examples{
library(labyrinth)

data("ncbi_info", package = "labyrinth")

}
\keyword{datasets}
